refactor(service): Os mudança de diretorio dos services de documents-to-send
