+++ src/app/edocs/providers/index.ts
@@ -1,1 +1,3 @@
 export * from './wizard-step';
+export * from './documents-to-send.service';
+export * from './documents-to-send.api.service';
