+++ src/app/edocs/state/index.ts
@@ -2,8 +2,7 @@ import { DocumentsApiService } from './documents.api.service';
 import { DocumentsQuery } from './documents.query';
 import { DocumentsService } from './documents.service';
 import { DocumentsStore } from './documents.store';
-import { DocumentsToSendApiService } from './documents-to-send.api.service';
-import { DocumentsToSendService } from './documents-to-send.service';
+import { DocumentsToSendApiService, DocumentsToSendService } from '../providers';
 import { DocumentsToSendQuery } from './documents-to-send.query';
 import { DocumentsToSendStore } from './documents-to-send.store';
